requirements.txt
shortuuid==1.0.11
python-dotenv==1.0.0
pandas==1.5.3